change getEmail() to show the flags also size and udate
I like the idea in your code where the functions can be chained, but I have found some issues.
When reading from my email account I noticed that all the emails were marked as read/seen! Could it be possible to add a comment to the readme file so that the mark_as_read parameter is set to true or preffereably change the default to false in line 144 or Reader.php (not in Pull Request)

Having fixed the above issue the unseen flag was still not set in your code I delved deeper to see that you used imap_rfc822_parse_headers, in line 890. Unfortunately this does not set: size, udate, recent, unseen, flagged, answered, deleted, draft. In order for these to be read from the server one must use imap_headerinfo and pass the stream and message number instead of uid.

However the changes to the imap spec, no longer RFC3501 and now RFC9051, means that the recent flag is obselete.

I notice that you use $this->stream() often which will has reconnect=true as deffault so I have added false to avoid multiple reconnections to the server. Your function stream tests if is_resource. Does this work if using php8.1 (I have not upgraded but I am using php8.0)